check if sys_path is dir first

+++ src/flay/bundle/package.py
@@ -116,7 +116,7 @@ def bundle_package(
     # look for {top_level_package}.libs dir and bundle it
     # i.e. the musllinux build of pydantic-core needs this
     for sys_path in sys.path:
-        if os.path.exists(sys_path):
+        if os.path.exists(sys_path) and os.path.isdir(sys_path):
             for dir_ in os.listdir(sys_path):
                 if dir_ == f"{top_level_package}.libs":
                     shutil.copy(f"{sys_path}/{dir_}", destination_path / dir_)
